Move the code from `ui` directory to `ui-common`
