--- ui-common/README.md
@@ -1,30 +1,0 @@
-# kubeshark-common
-
-> Made with create-react-library
-
-[![NPM](https://img.shields.io/npm/v/liraz-test.svg)](https://www.npmjs.com/package/liraz-test) [![JavaScript Style Guide](https://img.shields.io/badge/code_style-standard-brightgreen.svg)](https://standardjs.com)
-
-## Install
-
-```bash
-npm install --save @up9/kubeshark-common
-```
-
-## Usage
-
-```tsx
-import React, { Component } from 'react'
-
-import MyComponent from 'l@up9/kubeshark-common'
-import '@up9/kubeshark-common/dist/index.css'
-
-class Example extends Component {
-  render() {
-    return <MyComponent />
-  }
-}
-```
-
-## License
-
-MIT © [](https://github.com/)
