+++ ui-common/public/manifest.json
