+++ ui-common/src/components/AuthPresentation/AuthPresentation.sass
