+++ ui-common/src/components/TrafficViewer/TrafficViewer.tsx
@@ -5,8 +5,8 @@ import makeStyles from '@mui/styles/makeStyles';
 import TrafficViewerStyles from "./TrafficViewer.module.sass";
 import styles from '../EntriesList/EntriesList.module.sass';
 import { EntryDetailed } from "../EntryDetailed/EntryDetailed";
-import playIcon from 'assets/run.svg';
-import pauseIcon from 'assets/pause.svg';
+import playIcon from "./assets/run.svg";
+import pauseIcon from "./assets/pause.svg";
 import variables from '../../variables.module.scss';
 import { ToastContainer } from 'react-toastify';
 import { RecoilRoot, RecoilState, useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
