+++ ui-common/src/components/UI/FancyTextDisplay/FancyTextDisplay.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
-import duplicateImg from "assets/duplicate.svg";
+import duplicateImg from "./assets/duplicate.svg";
 import styles from './FancyTextDisplay.module.sass';
 
 interface Props {
