+++ ui-common/src/components/UI/SearchableDropdown/SearchableDropdown.tsx
@@ -4,7 +4,7 @@ import { Autocomplete } from "@mui/material";
 import { Checkbox, TextField } from "@mui/material";
 import CheckBoxOutlineBlankIcon from '@mui/icons-material/CheckBoxOutlineBlank';
 import CheckBoxIcon from '@mui/icons-material/CheckBox';
-import DefaultIconDown from "assets/DefaultIconDown.svg";
+import DefaultIconDown from "./assets/DefaultIconDown.svg";
 import styles from "./SearchableDropdown.module.sass";
 
 interface SearchableDropdownProps {
