+++ ui-common/src/consts.ts
