+++ ui-common/src/typings.d.ts
